Control addition of Questions to Topic
To begin implementing sorting of the list of questions, firstly control
how they are added. Here we prevent adding a question using the
append() method on the recentQuestions list by making the list
externally immutable, then adding an add method. This enables us to
control how Questions are added to the array, while still allowing
read access to the array.

Issue #1