Begin implementation of StackOverFlowManager
Create initial implementation of StackOverflowManager which initially
asks its communicator to search for questions on a topic.

A mock communicator is created, which is used to verify the search was
requested.

Note that the mock object is created in the test fixture, whereas the
communicator protocol is created within the application - which is
where it will be implemented.

Issue #5