Add error condition for question search
This uses the traditional NSError approach to dealing with errors.

Here the StackOverflowManager passes an error to the delegate.
However, it does not pass on the error it receives from the
Communicator - as this could be a low-level error (e.g., a DNS entry
was not found). The StackOverflowManager just reports there was an
error fetching the questions.

Issue #5