Robustness test case for nil underlying error
Need to add a robustness test case to ensure the delegate can handle an
error (due to nil questions) which has no underlying error. This
handles the case in the book where the
testQuestionsJSONIsPassedToQuestionBuilder crashes.

Note, we do not get a crash in Swift as we use a different error
handling mechanism. However, prior to this when using NSError we were
using optionals which prevented the program from crashing.

Issue #5