Add tests score property of Question
Note: The date field needs to be set to the same value before
comparing scores. Since the date is set on initialisation to the
current date/dime, any two Questions will have a different date value
(though possibly only microseconds apart). To ignore the date property
for testing the score property, need to ensure the dates are exactly
the same be explicitly setting them to the same value (in this case
the distantFuture value).

Note: For the second test, we ensure the questions are the same before
setting the score properties. This is to ensure only the score
properties impact the result of the following assertion.

Issue #2