Compare question arrays not just the count
Previously the test only compared the length of the arrays to
determine if they were the same. This is not sufficient as two arrays
could be the same length, but have different contents. To enable
comparing the arrays the Question type needs to conform to the
Equatable protocol. Now that has been achieved, we can update the
test to compare the arrays.

Issue #5