Cancel a task when it is completed
This test requires using a URLSession created with a delegate - since
we are calling the delegate method. We set this session to be the sut
session - to override the mock session in the setUp() method. However,
we use a mock data task since we want to know that the data task
cancel() method is called.