Change how error test functions
To enable injecting an error status code into the search query, have
used a StackOverflowCommunicator with introspection capability. That
is, the private method fetchContentsAtUrl() is overridden so an error
status code can be injected. This also requires overriding the
searchForQuestions() method - so the overridden fetchContentsAtUrl()
is used instead of the one from the original
StackOverflowCommunicator. This should be fine as long as the
serachForQuestions() methods are exactly the same.

Also, in the fetchContentsAtUrl() method, the delegate method
urlSession(_:task:didCompleteWithError) is called to pass the error to
the delegate - which is what we are attempting to test here.