Have communicator send data received to delegate
Prior to sending, the data is converted to a string.

The test injects data into the communicator by calling the delegate
method it implements urlSession(_:dataTask:didReceive).

We then call the searchForQuestions() method to ensure the FetchType
is set for the relevant query to the Stack Overflow API.

Since we are using the DataInsertingStackOverlfowCommunicator, we
bypass the fetchContentsAtUrl() and use this to just call the
urlSession(_:task:didCompleteWithError). This method then calls the
relevant method to send data to the delegate.

Note: We need to use DataInsertingStackOverflowCommunicator instead of
ErrorInsertingStackOverflowCommunicator - since the latter inserts and
error, and will bypass sending data to the delegate (sending an error
instead).

There are two things we are bypassing here:
 1. A call to StackOverflowCommunicator.fetchContentAtUrl()
 2. Using the data task created in this method and using a
    MockDataTask() instead.
The aim is to bypass the netwrorking for the unit test. This is OK, as
we just want to ensure we are sending the data to the
StackOverflowCommunicatorDelegate - which is the StackOverflowManager.

Issue #8